requirements.txt
absl-py==0.12.0
aiohttp-cors==0.7.0
aiohttp==3.7.4.post0
aioredis==1.3.1
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blessings==1.7
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
codecov==2.1.11
colorama==0.4.4
colorful==0.5.4
configparser==5.0.2
coverage==5.5
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dm-tree==0.1.5
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.5
gitpython==3.1.14
google-api-core==1.26.1
google-auth-oauthlib==0.4.3
google-auth==1.27.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
gputil==1.4.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
hiredis==1.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.7.2
iniconfig==1.1.1
ipykernel==6.5.1
ipython-genutils==0.2.0
ipython==7.29.0
ipywidgets==7.6.5
jedi==0.18.1
jinja2==3.0.3
jsonschema==3.2.0
jssenv==1.0.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaleido==0.2.1
keras-preprocessing==1.1.2
lz4==3.1.3
markdown==3.3.4
markupsafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
multidict==5.1.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus-context==0.1.2
opencensus==0.7.12
opencv-python-headless==4.3.0.36
opencv-python==4.5.1.48
opt-einsum==3.3.0
ortools==9.0.9048
packaging==20.9
pandas==1.2.3
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==7.2.0
pip==21.1.1
plotly==4.14.3
pluggy==0.13.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.22
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
py-spy==0.3.4
py==1.10.0
pyarrow==6.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pydeck==0.7.1
pyglet==1.5.0
pygments==2.10.0
pympler==0.9
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
pytz-deprecation-shim==0.1.0.post0
pytz==2021.1
pyyaml==5.4.1
pyzmq==22.3.0
ray==1.1.0
redis==3.5.3
requests-oauthlib==1.3.0
requests==2.25.1
retrying==1.3.3
rsa==4.7.2
scipy==1.6.1
send2trash==1.8.0
sentry-sdk==1.0.0
setproctitle==1.1.10
setuptools==56.0.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
streamlit==1.2.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard-plugin-wit==1.8.0
tensorboard==2.2.2
tensorboardx==2.1
tensorflow-estimator==2.2.0
tensorflow==2.2.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.4
validators==0.18.2
wandb==0.10.22
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1